// types/Register.ts
export interface Location {
  province: string;
  district: string;
  subdistrict: string;
  zipcode: string;
}